pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "AtomSculptor"
version = "0.0.1"
description = "Structure building agent team"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.26.0",
    "litellm>=1.82.0",
    "pyyaml>=6.0.3",
    "pymatgen>=2025.10.7",
    "ase>=3.26.0",
    "rdkit>=2025.9.6",
    "duckduckgo_search>=8.1.1",
    # code-graph-rag dependencies
    "loguru>=0.7.3",
    "pydantic-ai-slim[google,openai,vertexai]>=0.2.18",
    "pydantic-settings>=2.0.0",
    "pymgclient>=1.4.0",
    "python-dotenv>=1.1.0",
    "toml>=0.10.2",
    "tree-sitter-python>=0.23.6",
    "tree-sitter>=0.24.0",
    "watchdog>=6.0.0",
    "diff-match-patch>=20241021",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.5",
    "mypy>=1.11.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]

treesitter-full = [
    "tree-sitter-javascript>=0.23.1",
    "tree-sitter-typescript>=0.23.2",
    "tree-sitter-rust>=0.24.0",
    "tree-sitter-go>=0.23.4",
    "tree-sitter-scala>=0.24.0",
    "tree-sitter-java>=0.23.5",
    "tree-sitter-cpp>=0.23.0",
]

[tool.setuptools.packages.find]
where = [".", "code_graph_rag"]
include = [
    "agent_team", "agent_team.*",
    "sandbox", "sandbox.*",
    "code_graph_rag", "code_graph_rag.*",
    "codebase_rag", "codebase_rag.*",
]